Refactor distance.hpp and event_fetcher.hpp code structure: Move implementation details to details folder
